pyproject.toml
[tool.poetry]
name = "admissions"
version = "0.1.0"
description = "Admission system for Abakus"
authors = ["Webkom <webkom@abakus.no>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
django-environ = "^0.10.0"
django-extensions = "^3.2.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.0.0"
django-vite = "^2.1.0"
social-auth-app-django = "^5.2.0"
psycopg2-binary = "^2.9.6"
ipython = "^8.13.2"
structlog = "^23.1.0"
sentry-sdk = "^1.24.0"
uwsgi = "^2.0.21"
premailer = "^3.10.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2.0"
factory-boy = "^2.12.0"
tox = "^4.6.4"
isort = "^5.12.0"
black = "^23.7.0"
coverage = "^7.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.5.1"]
build-backend = "poetry.core.masonry.api"
